[fix] Compilation fixes related to shared collision system
